Agr2 100 backend implement farm based authorization middleware rbac (#35)
* protect routes with policy

* fix: Update GetAnimalByIdQueryHandlerTests to support new Farm Context validation

* refactor: Use AutoMocker in GetAnimalByIdQueryHandlerTests for cleaner setup

* add core E2E tests

* reformat code

* update testcontainers

* update dependencies

* apply IDOR suggestions

* use existing handle system error

* check farm when updating resources

* apply suggestions

* remove TODO

* improve validations in controllers

* use automoq in all unit tests

* secure movement routes

* format code

+++ src/AgroLink.Application/Features/Paddocks/Commands/Delete/DeletePaddockCommand.cs
@@ -1,3 +1,5 @@
+using AgroLink.Application.Common.Exceptions;
+using AgroLink.Application.Interfaces;
 using AgroLink.Domain.Interfaces;
 using MediatR;
 
@@ -7,6 +9,7 @@ public record DeletePaddockCommand(int Id) : IRequest;
 
 public class DeletePaddockCommandHandler(
     IPaddockRepository paddockRepository,
+    ICurrentUserService currentUserService,
     IUnitOfWork unitOfWork
 ) : IRequestHandler<DeletePaddockCommand>
 {
@@ -16,6 +19,14 @@ public class DeletePaddockCommandHandler(
         if (paddock == null)
         {
             throw new ArgumentException("Paddock not found");
+        }
+
+        if (
+            currentUserService.CurrentFarmId.HasValue
+            && paddock.FarmId != currentUserService.CurrentFarmId.Value
+        )
+        {
+            throw new ForbiddenAccessException("You do not have access to this paddock");
         }
 
         paddockRepository.Remove(paddock);
